#4739 move sdk to its own module within kepn/keptn
