+++ remediation-service/handler/get_action_event_handler_test.go
@@ -8,8 +8,8 @@ import (
 	"github.com/keptn/go-utils/pkg/lib/v0_1_4"
 	keptnv2 "github.com/keptn/go-utils/pkg/lib/v0_2_0"
 	"github.com/keptn/keptn/remediation-service/handler"
-	"github.com/keptn/keptn/remediation-service/internal/sdk"
-	"github.com/keptn/keptn/remediation-service/internal/sdk/fake"
+	sdk "github.com/keptn/keptn/sdk/pkg"
+	"github.com/keptn/keptn/sdk/pkg/fake"
 	"github.com/stretchr/testify/require"
 	"io/ioutil"
 	"log"
