+++ sdk/go.mod
@@ -1,0 +1,13 @@
+module github.com/keptn/keptn/sdk
+
+go 1.16
+
+require (
+	github.com/cloudevents/sdk-go/v2 v2.4.1
+	github.com/google/uuid v1.3.0
+	github.com/kelseyhightower/envconfig v1.4.0
+	github.com/keptn/go-utils v0.8.5
+	github.com/sirupsen/logrus v1.8.1
+	github.com/stretchr/testify v1.7.0
+	golang.org/x/sys v0.0.0-20200202164722-d101bd2416d5 // indirect
+)
