+++ sdk/pkg/events.go
@@ -1,4 +1,4 @@
-package sdk
+package pkg
 
 import (
 	"context"
