+++ sdk/pkg/fake/event_sender_mock.go
@@ -5,26 +5,26 @@ package fake
 
 import (
 	"github.com/cloudevents/sdk-go/v2/event"
-	"github.com/keptn/keptn/remediation-service/internal/sdk"
+	"github.com/keptn/keptn/sdk/pkg"
 	"sync"
 )
 
-// Ensure, that EventSenderMock does implement sdk.EventSender.
+// Ensure, that EventSenderMock does implement pkg.EventSender.
 // If this is not the case, regenerate this file with moq.
-var _ sdk.EventSender = &EventSenderMock{}
+var _ pkg.EventSender = &EventSenderMock{}
 
-// EventSenderMock is a mock implementation of sdk.EventSender.
+// EventSenderMock is a mock implementation of pkg.EventSender.
 //
 // 	func TestSomethingThatUsesEventSender(t *testing.T) {
 //
-// 		// make and configure a mocked sdk.EventSender
+// 		// make and configure a mocked pkg.EventSender
 // 		mockedEventSender := &EventSenderMock{
 // 			SendEventFunc: func(eventMoqParam event.Event) error {
 // 				panic("mock out the SendEvent method")
 // 			},
 // 		}
 //
-// 		// use mockedEventSender in code that requires sdk.EventSender
+// 		// use mockedEventSender in code that requires pkg.EventSender
 // 		// and then make assertions.
 //
 // 	}
