+++ webhook-service/go.mod
@@ -1,0 +1,3 @@
+module github.com/keptn/keptn/webhook-service
+
+go 1.16
